refactor: UserInputController 추가
