+++ src/main/kotlin/controller/GameController.kt
@@ -4,7 +4,6 @@ import org.example.config.AppConfig
 import org.example.domain.LottoGame
 import org.example.domain.LottoTicket
 import org.example.domain.PrizeCategory
-import org.example.domain.WinningLotto
 import org.example.service.LottoService
 import org.example.view.InputView
 import org.example.view.OutputView
@@ -13,20 +12,22 @@ class GameController(
     private val inputView: InputView,
     private val outputView: OutputView,
     private val lottoService: LottoService,
+    private val userInputController: UserInputController,
 ) {
     constructor(config: AppConfig) : this(
         config.inputView,
         config.outputView,
         config.lottoService,
+        config.userInputController,
     )
 
     fun start(remainMoney: Int = 0): Int {
         // 구매금액 입력
-        val purchaseMoney = inputPurchaseAmount(remainMoney)
+        val purchaseMoney = userInputController.inputPurchaseAmount(remainMoney)
         // 수동으로 구매할 티켓 수 입력
-        val manualTicketCount = inputManualTicketCount(purchaseMoney)
+        val manualTicketCount = userInputController.inputManualTicketCount(purchaseMoney)
         // 수동으로 구매할 티켓 번호 입력
-        val manualTickets = inputManualTickets(manualTicketCount)
+        val manualTickets = userInputController.inputManualTickets(manualTicketCount)
         // 자동 생성된 티켓 출력
         val autoLottoTickets = generateAutoLotto(purchaseMoney - manualTicketCount)
         // 전체 로또 티켓 출력
@@ -35,7 +36,7 @@ class GameController(
         val winningMode = selectWinningMode()
         // 당첨 번호 입력
         // 보너스 번호 입력
-        val winningLotto = inputWinningNumbers(winningMode)
+        val winningLotto = userInputController.inputWinningNumbers(winningMode)
         // 결과 출력
         val lottoGameResult =
             LottoGame(
@@ -63,33 +64,6 @@ class GameController(
         outputView.displayTotalEarning(lottoService.calculateTotalEarningPercentage(purchaseMoney, totalResult))
     }
 
-    private fun inputManualTickets(manualTicketCount: Int): List<LottoTicket> {
-        if (manualTicketCount == 0) return emptyList()
-
-        try {
-            return inputView.inputManualTickets(manualTicketCount).map { lottoService.manualLottoGenerator(it) }
-                .toList()
-        } catch (e: IllegalArgumentException) {
-            outputView.displayInvalidInput(e.message)
-            return inputManualTickets(manualTicketCount)
-        }
-    }
-
-    private fun inputWinningNumbers(winningMode: String): WinningLotto {
-        if (winningMode == "A") {
-            return lottoService.autoWinningLottoGenerator().also { outputView.displayAutomaticWinningNumbers(it) }
-        }
-
-        try {
-            val winningLotto = inputView.inputWinningNumbers()
-            val bonusNumber = inputView.inputBonusNumber()
-            return WinningLotto(winningLotto, bonusNumber)
-        } catch (e: IllegalArgumentException) {
-            outputView.displayInvalidInput(e.message)
-            return inputWinningNumbers(winningMode)
-        }
-    }
-
     private fun printAllLottoTickets(
         manualTickets: List<LottoTicket>,
         autoLottoTickets: List<LottoTicket>,
@@ -106,28 +80,5 @@ class GameController(
         val autoTickets = lottoService.autoLottoGenerator(autoLottoTickets)
         outputView.displayAutoLottoTicket(autoTickets)
         return autoTickets
-    }
-
-    private fun inputManualTicketCount(purchaseMoney: Int): Int {
-        try {
-            return inputView.inputManualTicketCount(purchaseMoney)
-        } catch (e: IllegalArgumentException) {
-            outputView.displayInvalidInput(e.message)
-            return inputManualTicketCount(purchaseMoney)
-        }
-    }
-
-    private fun inputPurchaseAmount(remainMoney: Int): Int {
-        if (remainMoney != 0) {
-            return remainMoney
-        }
-
-        try {
-            val purchase = inputView.inputPurchaseAmount()
-            return purchase
-        } catch (e: IllegalArgumentException) {
-            outputView.displayInvalidInput(e.message)
-            return inputPurchaseAmount(remainMoney)
-        }
     }
 }
